feat: User와 Article 관계 맵핑
- Article 수정 및 삭제시 권한 확인하는 로직 추가
- 세션에 LoginUser 대신 User 저장하도록 변경

+++ src/main/java/techcourse/myblog/controller/ArticleController.java
@@ -3,8 +3,13 @@ package techcourse.myblog.controller;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import techcourse.myblog.domain.Article;
+import techcourse.myblog.domain.User;
 import techcourse.myblog.dto.ArticleDto;
+import techcourse.myblog.exception.InvalidAuthorException;
 import techcourse.myblog.service.ArticleService;
+
+import javax.servlet.http.HttpSession;
 
 @Controller
 public class ArticleController {
@@ -26,8 +31,9 @@ public class ArticleController {
     }
 
     @PostMapping("/articles")
-    public String writeArticle(ArticleDto articleDto) {
-        long articleId = articleService.save(articleDto).getArticleId();
+    public String writeArticle(ArticleDto articleDto, HttpSession httpSession) {
+        User user = (User) httpSession.getAttribute("user");
+        long articleId = articleService.save(articleDto, user).getArticleId();
         return "redirect:/articles/" + articleId;
     }
 
@@ -39,21 +45,33 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/{articleId}/edit")
-    public String showEditForm(@PathVariable long articleId, Model model) {
+    public String showEditForm(@PathVariable long articleId, Model model, HttpSession httpSession) {
+        User user = (User) httpSession.getAttribute("user");
+        Article article = articleService.getArticleById(articleId);
+        article.checkCorrespondingAuthor(user);
         ArticleDto articleDto = articleService.getArticleDtoById(articleId);
         model.addAttribute("articleDto", articleDto);
         return "article-edit";
     }
 
     @PutMapping("/articles/{articleId}")
-    public String updateArticle(@PathVariable long articleId, ArticleDto articleDto) {
-        articleService.update(articleDto);
+    public String updateArticle(@PathVariable long articleId, ArticleDto articleDto, HttpSession httpSession) {
+        User user = (User) httpSession.getAttribute("user");
+        articleService.update(articleDto, user);
         return "redirect:/articles/" + articleId;
     }
 
     @DeleteMapping("/articles/{articleId}")
-    public String deleteArticle(@PathVariable long articleId) {
+    public String deleteArticle(@PathVariable long articleId, HttpSession httpSession) {
+        User user = (User) httpSession.getAttribute("user");
+        Article article = articleService.getArticleById(articleId);
+        article.checkCorrespondingAuthor(user);
         articleService.deleteById(articleId);
+        return "redirect:/";
+    }
+
+    @ExceptionHandler(InvalidAuthorException.class)
+    public String handleInvalidAuthorException(InvalidAuthorException e) {
         return "redirect:/";
     }
 }
